requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.2.4
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
boto3==1.37.4
botocore==1.37.4
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dill==0.3.9
Django==4.2.19
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.0.0
gunicorn==23.0.0
isort==5.13.2
jmespath==1.0.1
kombu==5.4.2
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
